+++ components/StuListPage/StuInfo/style.ts
@@ -1,0 +1,24 @@
+import styled from '@emotion/styled'
+import * as MyPage from '@/components/MyPage/style'
+
+export const ProfileLayer = styled.div`
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  gap: 60px;
+  small {
+    color: rgba(255, 255, 255, 0.31);
+  }
+`
+
+export const ProfileBox = styled(MyPage.ProfileBox)``
+
+export const ProfileContent = styled(MyPage.ProfileContent)``
+
+export const ProfileImg = styled(MyPage.ProfileImg)``
+
+export const ClubBox = styled(MyPage.ClubBox)``
+
+export const ClubContainer = styled(MyPage.ClubContainer)``
+
+export const ClubType = styled(MyPage.ClubType)``
